+++ src/hooks/useForm.js
@@ -1,0 +1,67 @@
+import { useReducer } from 'react';
+
+const actionTypes = {
+  inputChange: 'INPUT CHANGE',
+  clearValues: 'CLEAR VALUES',
+  consentToggle: 'CONSENT TOGGLE',
+  throwError: 'THROW ERROR',
+};
+
+const reducer = (state, action) => {
+  switch (action.type) {
+    case actionTypes.inputChange:
+      return {
+        ...state,
+        [action.field]: action.value,
+      };
+    case actionTypes.clearValues:
+      return {
+        ...action.initialValues,
+      };
+    case actionTypes.consentToggle:
+      console.log('zmiana zgody');
+      return {
+        ...state,
+        consent: !state.consent,
+      };
+    case 'THROW ERROR':
+      return {
+        ...state,
+        error: [action.errorValue],
+      };
+    default:
+      return state;
+  }
+};
+
+export const useForm = (initialValues) => {
+  const [formValues, dispatch] = useReducer(reducer, initialValues);
+
+  const handleInputChange = (e) => {
+    dispatch({
+      type: 'INPUT CHANGE',
+      field: e.target.name,
+      value: e.target.value,
+    });
+  };
+
+  const handleClearForm = (initialValues) => {
+    dispatch({ type: actionTypes.clearValues }, initialValues);
+  };
+
+  const handleThrowError = (errorMessage) => {
+    dispatch({ type: actionTypes.throwError, errorValue: errorMessage });
+  };
+
+  const handleToggleConsent = () => {
+    dispatch({ type: actionTypes.consentToggle });
+  };
+
+  return {
+    formValues,
+    handleInputChange,
+    handleClearForm,
+    handleThrowError,
+    handleToggleConsent,
+  };
+};
